pin-mux: Rename sol-pin-mux-impl.h
To sol-pin-mux-modules.h to avoid misinterpreting it as an external
header wrongly exposed.
