+++ src/modules/pin-mux/intel-galileo-rev-d/intel-galileo-rev-d.c
@@ -31,7 +31,7 @@
  */
 
 #include "sol-util.h"
-#include "sol-pin-mux-impl.h"
+#include "sol-pin-mux-modules.h"
 
 // =============================================================================
 // Galileo Gen1 Multiplexer Description
